scraper/requirements.txt
selenium==4.11.2
supabase==1.0.3
python-dotenv==1.0.0
schedule==1.2.0
webdriver-manager==4.0.1
python-dateutil==2.8.2
flask==3.0.0

